requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
faiss-cpu==1.11.0.post1
filelock==3.18.0
fsspec==2025.7.0
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pillow==11.3.0
PyMuPDF==1.26.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.3
typing_extensions==4.14.1
urllib3==2.5.0
